Factor out retrieval of URL metadata into lz-http
