+++ Cargo.toml
@@ -11,6 +11,7 @@ members = [
   "src/lz-cli",
   "src/lz-openapi",
   "src/lz-ui",
+  "src/lz-http",
 ]
 
 [workspace.package]
@@ -73,3 +74,4 @@ web-sys = "0.3.69"
 tracing-web = "0.1.3"
 yew-router = "0.18.0"
 yew-autocomplete = "0.3.1"
+itertools = "0.12.1"
