+++ src/lz-http/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "lz-http"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+reqwest.workspace = true
+scraper.workspace = true
+thiserror.workspace = true
+url.workspace = true
