+++ src/lz-ui/Cargo.toml
@@ -20,4 +20,4 @@ tracing-web = { workspace = true }
 tracing-subscriber.workspace = true
 yew-router = { workspace = true }
 yew-autocomplete = { workspace = true }
-itertools = "0.12.1"
+itertools = { workspace = true }
